Remove validation err list; add "_get_child_position" doc string
The validation err list is meaningful during tree-based validation. The
"_get_child_position" function is not part of the validation workflow,
but rather is used only to determine the position of a child node within
a parent node and the MetapypeRuleError only occurs if the child is not
permitted in the parent.